DELFOR Integration
YYY(trading partner) would send DELFORs in EDIFACT format to XXX(client) via an FTP connection(FTP V2). We would then route the files for deenveloping in the trading partner step/shape. 
We then split the documents such that they each documents contains only one Line item. XXX has a requirement that before the documents are imported in SAP, it needs to have only one Line item per transaction, or else these will not be processed.
The next step is to route the inbound DELFOR documents to the respective map which will process DELFORs containing SCC+1 which is immediate requirement records and SCC+4 which are forecast records to the DELFOR_IX_TP_MAP_V7 map step/shape and DELFORS with only SCC+4 that is having only Forecast records to the DELFOR_IX_TP_MAP_V8 map step/shape.
The logic for filtering the documents in the digital step is by checking if the profile element value of QTY with SCC qualifier 1 is null or not.

Once the documents have been filtered to their respective maps, these are mapped to the SAP IDoc structure and sent to the Client SAP system. 
(Note: we had configured an SFTP client which connects to the directory on the client server. Since the client was in the process changing their transmission protocol, we had to disconnect the flow and is in the process of connectivity testing).